remove all unneeded from workspace
