+++ src/client/testing/testController/workspaceTestAdapter.ts
@@ -1,19 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import * as path from 'path';
 import * as util from 'util';
-import {
-    CancellationToken,
-    Position,
-    Range,
-    TestController,
-    TestItem,
-    TestMessage,
-    TestRun,
-    Uri,
-    Location,
-} from 'vscode';
+import { CancellationToken, TestController, TestItem, TestMessage, TestRun, Uri, Location } from 'vscode';
 import { splitLines } from '../../common/stringUtils';
 import { createDeferred, Deferred } from '../../common/utils/async';
 import { Testing } from '../../common/utils/localize';
@@ -21,23 +10,11 @@ import { traceError, traceVerbose } from '../../logging';
 import { sendTelemetryEvent } from '../../telemetry';
 import { EventName } from '../../telemetry/constants';
 import { TestProvider } from '../types';
-import {
-    clearAllChildren,
-    createErrorTestItem,
-    DebugTestTag,
-    ErrorTestItemOptions,
-    getTestCaseNodes,
-    RunTestTag,
-} from './common/testItemUtilities';
-import {
-    DiscoveredTestItem,
-    DiscoveredTestNode,
-    DiscoveredTestType,
-    ITestDiscoveryAdapter,
-    ITestExecutionAdapter,
-} from './common/types';
+import { clearAllChildren, createErrorTestItem, getTestCaseNodes } from './common/testItemUtilities';
+import { ITestDiscoveryAdapter, ITestExecutionAdapter } from './common/types';
 import { fixLogLines } from './common/utils';
 import { IPythonExecutionFactory } from '../../common/process/types';
+import { buildErrorNodeOptions } from './common/resultResolver';
 
 /**
  * This class exposes a test-provider-agnostic way of discovering tests.
@@ -291,8 +268,6 @@ export class WorkspaceTestAdapter {
     ): Promise<void> {
         sendTelemetryEvent(EventName.UNITTEST_DISCOVERING, undefined, { tool: this.testProvider });
 
-        const workspacePath = this.workspaceUri.fsPath;
-
         // Discovery is expensive. If it is already running, use the existing promise.
         if (this.discovering) {
             return this.discovering.promise;
@@ -301,13 +276,12 @@ export class WorkspaceTestAdapter {
         const deferred = createDeferred<void>();
         this.discovering = deferred;
 
-        let rawTestData;
         try {
             // ** execution factory only defined for new rewrite way
             if (executionFactory !== undefined) {
-                rawTestData = await this.discoveryAdapter.discoverTests(this.workspaceUri, executionFactory);
+                await this.discoveryAdapter.discoverTests(this.workspaceUri, executionFactory);
             } else {
-                rawTestData = await this.discoveryAdapter.discoverTests(this.workspaceUri);
+                await this.discoveryAdapter.discoverTests(this.workspaceUri);
             }
             deferred.resolve();
         } catch (ex) {
@@ -334,147 +308,6 @@ export class WorkspaceTestAdapter {
             // we don't need the deferred promise anymore.
             this.discovering = undefined;
         }
-
-        if (!rawTestData) {
-            // No test data is available
-            return Promise.resolve();
-        }
-
-        // Check if there were any errors in the discovery process.
-        if (rawTestData.status === 'error') {
-            const testingErrorConst =
-                this.testProvider === 'pytest' ? Testing.errorPytestDiscovery : Testing.errorUnittestDiscovery;
-            const { errors } = rawTestData;
-            traceError(testingErrorConst, '\r\n', errors!.join('\r\n\r\n'));
-
-            let errorNode = testController.items.get(`DiscoveryError:${workspacePath}`);
-            const message = util.format(
-                `${testingErrorConst} ${Testing.seePythonOutput}\r\n`,
-                errors!.join('\r\n\r\n'),
-            );
-
-            if (errorNode === undefined) {
-                const options = buildErrorNodeOptions(this.workspaceUri, message, this.testProvider);
-                errorNode = createErrorTestItem(testController, options);
-                testController.items.add(errorNode);
-            }
-            errorNode.error = message;
-        } else {
-            // Remove the error node if necessary,
-            // then parse and insert test data.
-            testController.items.delete(`DiscoveryError:${workspacePath}`);
-
-            // Wrap the data under a root node named after the test provider.
-            const wrappedTests = rawTestData.tests;
-
-            // If we are in a multiroot workspace scenario, wrap the current folder's test result in a tree under the overall root + the current folder name.
-            let rootPath = workspacePath;
-            let childrenRootPath = rootPath;
-            let childrenRootName = path.basename(rootPath);
-
-            if (isMultiroot) {
-                rootPath = workspaceFilePath!;
-                childrenRootPath = workspacePath;
-                childrenRootName = path.basename(workspacePath);
-            }
-
-            const children = [
-                {
-                    path: childrenRootPath,
-                    name: childrenRootName,
-                    type_: 'folder' as DiscoveredTestType,
-                    id_: childrenRootPath,
-                    children: wrappedTests ? [wrappedTests] : [],
-                },
-            ];
-
-            // Update the raw test data with the wrapped data.
-            rawTestData.tests = {
-                path: rootPath,
-                name: this.testProvider,
-                type_: 'folder',
-                id_: rootPath,
-                children,
-            };
-
-            if (rawTestData.tests) {
-                // If the test root for this folder exists: Workspace refresh, update its children.
-                // Otherwise, it is a freshly discovered workspace, and we need to create a new test root and populate the test tree.
-                populateTestTree(testController, rawTestData.tests, undefined, this, token);
-            } else {
-                // Delete everything from the test controller.
-                testController.items.replace([]);
-            }
-        }
-
-        sendTelemetryEvent(EventName.UNITTEST_DISCOVERY_DONE, undefined, { tool: this.testProvider, failed: false });
         return Promise.resolve();
     }
-}
-
-function isTestItem(test: DiscoveredTestNode | DiscoveredTestItem): test is DiscoveredTestItem {
-    return test.type_ === 'test';
-}
-
-// had to switch the order of the original parameter since required param cannot follow optional.
-function populateTestTree(
-    testController: TestController,
-    testTreeData: DiscoveredTestNode,
-    testRoot: TestItem | undefined,
-    wstAdapter: WorkspaceTestAdapter,
-    token?: CancellationToken,
-): void {
-    // If testRoot is undefined, use the info of the root item of testTreeData to create a test item, and append it to the test controller.
-    if (!testRoot) {
-        testRoot = testController.createTestItem(testTreeData.path, testTreeData.name, Uri.file(testTreeData.path));
-
-        testRoot.canResolveChildren = true;
-        testRoot.tags = [RunTestTag, DebugTestTag];
-
-        testController.items.add(testRoot);
-    }
-
-    // Recursively populate the tree with test data.
-    testTreeData.children.forEach((child) => {
-        if (!token?.isCancellationRequested) {
-            if (isTestItem(child)) {
-                const testItem = testController.createTestItem(child.id_, child.name, Uri.file(child.path));
-                testItem.tags = [RunTestTag, DebugTestTag];
-
-                const range = new Range(
-                    new Position(Number(child.lineno) - 1, 0),
-                    new Position(Number(child.lineno), 0),
-                );
-                testItem.canResolveChildren = false;
-                testItem.range = range;
-                testItem.tags = [RunTestTag, DebugTestTag];
-
-                testRoot!.children.add(testItem);
-                // add to our map
-                wstAdapter.runIdToTestItem.set(child.runID, testItem);
-                wstAdapter.runIdToVSid.set(child.runID, child.id_);
-                wstAdapter.vsIdToRunId.set(child.id_, child.runID);
-            } else {
-                let node = testController.items.get(child.path);
-
-                if (!node) {
-                    node = testController.createTestItem(child.id_, child.name, Uri.file(child.path));
-
-                    node.canResolveChildren = true;
-                    node.tags = [RunTestTag, DebugTestTag];
-                    testRoot!.children.add(node);
-                }
-                populateTestTree(testController, child, node, wstAdapter, token);
-            }
-        }
-    });
-}
-
-function buildErrorNodeOptions(uri: Uri, message: string, testType: string): ErrorTestItemOptions {
-    const labelText = testType === 'pytest' ? 'Pytest Discovery Error' : 'Unittest Discovery Error';
-    return {
-        id: `DiscoveryError:${uri.fsPath}`,
-        label: `${labelText} [${path.basename(uri.fsPath)}]`,
-        error: message,
-    };
 }
